requirements.txt
dash>=2.0.0
dash-bootstrap-components>=1.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
hdbscan>=0.8.27
python-dotenv>=0.19.0
gower>=0.1.0
numpy~=1.24.3
tensorflow-intel==2.15.0
sentence-transformers==2.2.2
transformers==4.30.2
tokenizers==0.13.3
huggingface-hub==0.16.4

pip~=25.1.1
wheel~=0.41.2
protobuf~=4.25.7
pillow~=11.2.1
Jinja2~=3.1.6
rsa~=4.9.1
pyasn1~=0.6.1
absl-py~=2.2.2
Markdown~=3.8
Flask~=3.0.3
importlib_metadata~=8.6.1
requests~=2.32.3
typing_extensions~=4.13.2
nest-asyncio~=1.6.0
Werkzeug~=3.0.6
retrying~=1.3.4
PyYAML~=6.0.2
plotly~=6.0.1
gast~=0.4.0
grpcio~=1.71.0
h5py~=3.13.0
pytz~=2025.2
rich~=14.0.0
keras~=2.15.0
colorama~=0.4.6
numba~=0.61.2
matplotlib~=3.10.1
umap~=0.1.1
umap-learn~=0.5.7
joblib~=1.4.2
scipy~=1.15.2
pynndescent~=0.5.13
tqdm~=4.67.1
zipp~=3.21.0
libclang~=18.1.1
MarkupSafe~=3.0.2
click~=8.1.8
blinker~=1.9.0
itsdangerous~=2.2.0
six~=1.17.0
wrapt~=1.14.1
llvmlite~=0.44.0
Pygments~=2.19.1
setuptools~=68.2.0
charset-normalizer~=3.4.1
google-pasta~=0.2.0
threadpoolctl~=3.6.0
torch~=2.7.0
mpmath~=1.3.0
sympy~=1.14.0
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=25.0
openpyxl~=3.1.5
XlsxWriter~=3.2.3
python-dateutil~=2.9.0.post0
narwhals~=1.36.0
oauthlib~=3.2.2
idna~=3.10
termcolor~=3.0.1
opt_einsum~=3.4.0
pyparsing~=3.2.3
tensorboard~=2.15.2
flatbuffers~=25.2.10
astunparse~=1.6.3
fsspec~=2025.3.2
google-auth-oauthlib~=1.2.2
requests-oauthlib~=2.0.0
kmodes~=0.12.2
wordcloud~=1.9.4
Flask-Caching~=2.3.1
nltk~=3.9.1